[xml-parse] Change `findall()` argument to str
Previously `findall()` accepted a DocumentType enum as one of its
arguments.  This was originally done because it was believed that there
was complete overlap in rfc-index.xml between entry types and
`documentId`.  This is not the case, however;  "rfc-not-issued-entry" is
a valid element under rfc-index but is not a valid document type.