Fix XML parsing and DB insertion of keywords
The name of the association table in ietf.sql.rfc is renamed to
"rfc_keyword" to be more descriptive.  Arguments have been added to the
column constructors for to explicitly declare a composite primary key
composed of `rfc_id` and `keyword_id`.

The `unique` attribute is applied to the `word` element of the keyword
table.  This necessitates the addition of an `_add_keyword()` function
to query the DB session for the prior existence of a keyword before
attempting to creating a new DB entry for the keyword.

Several `kw` entries in rfc-index.xml have leading whitespace.
`strip()` is called on the content of `kw` before returning the string.